more code

+++ Concepts/LeetCodeMath/Easy/171.cs
@@ -11,11 +11,10 @@ namespace LeetCodeMath.Easy
         /// <summary>
         //! Intuituon is converting base-26 number system to base-10 number system.
         /// </summary>
-        /// <param name="s"></param>
-        /// <returns></returns>
+
         public int TitleToNumber(string s)
         {
-            double result = 0;
+            int result = 0;
             int power = 0;
             for (int i = s.Length - 1; i >= 0; --i)
             {
@@ -24,9 +23,10 @@ namespace LeetCodeMath.Easy
                 //! Adding 1 will convert it into 1--26 number system
                 //! Another way to think is there is no 0 in our int system as it starts with one
                 ++number;
-                result = result + (int)((number - 'A') * Math.Pow(26, power++));
+                result = result + (number - 'A') *(int) Math.Pow(26, power++);
             }
-            return (int)result;
+
+            return result;
 
         }
